[REFACTOR]폴더명 변경 & db접근 모듈 생성(uniunity_db)
-교과서랑 똑같이 바꿈
- app.js->main.js
- routers->controllers
- uniunity_db모듈은 app/config/db.js
