--- app/config/config.js
@@ -1,18 +1,0 @@
-require('dotenv').config();
-module.exports={
-  development:{
-    username:process.env.DB_USER,
-    password:process.env.DB_PW,
-    database:process.env.DB_NAME,
-    host:process.env.DB_HOST,
-    dialect:"mysql"
-  },
-  production:{
-    username:process.env.DB_USER,
-    password:process.env.DB_PW,
-    database:process.env.DB_NAME,
-    host:process.env.DB_HOST,
-    dialect:"mysql",
-    logging:false
-  }
-}
