+++ app/main.js
@@ -1,26 +1,4 @@
 "use strict"
-//db
-require('dotenv').config();
-const mysql = require('mysql2/promise');
-
-let test=async()=>{
-    const db=mysql.createPool({
-        host:process.env.DB_HOST,
-        user:process.env.DB_USER,
-        password:process.env.DB_PW,
-        port:process.env.DB_PORT,
-        database:process.env.DB_NAME,
-        waitForConnections:true,
-        insecureAuth:true
-    });
-
-    let sql='SELECT * FROM subscriber';
-    let [rows,fields] =await db.query(sql);
-    console.log(rows);
-};
-test();
-
-
 
 // 모듈
 const express = require("express");
@@ -28,7 +6,9 @@ const bodyParser =require("body-parser");
 const app = express();
 
 // 라우팅
-const home = require("./src/routes/home");
+const home = require("./src/controllers/home");
+//db
+const uniunity_db = require("./config/db");
 
 // 앱 셋팅
 app.set("views","./src/views");
@@ -40,5 +20,8 @@ app.use(bodyParser.json());
 //URL을 통해 전달되는 데이터에 한글, 공백 등과 같은 문자가 포함될 경우 제대로 인식되지 않는 문제 해결
 app.use(bodyParser.urlencoded({extended:true}));
 app.use("/",home); //use -> 미들 웨어를 등록해주는 메서드
+
+
+uniunity_db();
 
 module.exports =app;
